validation

+++ webapp/app.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template, request, url_for, flash
+from flask import Flask, render_template, request, url_for, flash, session
 from werkzeug.utils import redirect
 from flask_mysqldb import MySQL
 
@@ -31,6 +31,7 @@ def login():
         cur.close()
         if data:
             flash("Logged in Successfully")
+            session['isAuth'] = True
             return redirect(url_for('logs'))
         else:
             flash("Invalid credentials")
@@ -42,34 +43,54 @@ def Loginp():
     return render_template('login.html')
 
 
+@app.route('/logout')
+def Logout():
+    session['isAuth'] = False
+    return redirect(url_for('Loginp'))
+
+
 @app.route('/employees')
 def Employees():
-    cur = mysql.connection.cursor()
-    cur.execute("SELECT * FROM employees")
-    data = cur.fetchall()
-    cur.close()
-
-    return render_template('employees.html', employees=data)
+    isAuthentic = session.get('isAuth', None)
+    if isAuthentic:
+        cur = mysql.connection.cursor()
+        cur.execute("SELECT * FROM employees")
+        data = cur.fetchall()
+        cur.close()
+        return render_template('employees.html', employees=data)
+    else:
+        flash("Login to access this page")
+        return redirect(url_for('Loginp'))
 
 
 @app.route('/edit/<string:id_data>')
 def Edit(id_data):
-    cur = mysql.connection.cursor()
-    cur.execute("SELECT * FROM employees WHERE id=%s", (id_data,))
-    data = cur.fetchall()
-    cur.close()
-
-    return render_template('edit.html', row=data)
+    isAuthentic = session.get('isAuth', None)
+    if isAuthentic:
+        cur = mysql.connection.cursor()
+        cur.execute("SELECT * FROM employees WHERE id=%s", (id_data,))
+        data = cur.fetchall()
+        cur.close()
+        return render_template('edit.html', row=data)
+    else:
+        flash("Login to access this page")
+        return redirect(url_for('Loginp'))
 
 
 @app.route('/logs')
 def logs():
-    cur = mysql.connection.cursor()
-    cur.execute("SELECT * FROM logs")
-    data = cur.fetchall()
-    cur.close()
+    isAuthentic = session.get('isAuth', None)
+    print(isAuthentic)
+    if isAuthentic:
+        cur = mysql.connection.cursor()
+        cur.execute("SELECT * FROM logs")
+        data = cur.fetchall()
+        cur.close()
 
-    return render_template('logs.html', logs=data)
+        return render_template('logs.html', logs=data)
+    else:
+        flash("Login to access this page")
+        return redirect(url_for('Loginp'))
 
 
 @app.route('/insert', methods=['POST'])
@@ -88,28 +109,38 @@ def insert():
 
 @app.route('/delete/<string:id_data>', methods=['GET'])
 def delete(id_data):
-    flash("Record Has Been Deleted Successfully")
-    cur = mysql.connection.cursor()
-    cur.execute("DELETE FROM employees WHERE id=%s", (id_data,))
-    mysql.connection.commit()
-    return redirect(url_for('Employees'))
+    isAuthentic = session.get('isAuth', None)
+    if isAuthentic:
+        flash("Record Has Been Deleted Successfully")
+        cur = mysql.connection.cursor()
+        cur.execute("DELETE FROM employees WHERE id=%s", (id_data,))
+        mysql.connection.commit()
+        return redirect(url_for('Employees'))
+    else:
+        flash("Login to access this page")
+        return redirect(url_for('Loginp'))
 
 
 @app.route('/update', methods=['GET', 'POST'])
 def update():
-    if request.method == 'POST':
-        id_data = request.form['id']
-        name = request.form['name']
-        email = request.form['email']
-        department = request.form['department']
-        cur = mysql.connection.cursor()
-        cur.execute("UPDATE employees SET name=%s, email=%s, department=%s WHERE id=%s",
-                    (name, email, department, id_data))
-        flash("Data Updated Successfully")
-        return redirect(url_for('Employees'))
+    isAuthentic = session.get('isAuth', None)
+    if isAuthentic:
+        if request.method == 'POST':
+            id_data = request.form['id']
+            name = request.form['name']
+            email = request.form['email']
+            department = request.form['department']
+            cur = mysql.connection.cursor()
+            cur.execute("UPDATE employees SET name=%s, email=%s, department=%s WHERE id=%s",
+                        (name, email, department, id_data))
+            flash("Data Updated Successfully")
+            return redirect(url_for('Employees'))
+        else:
+            flash("Data Updated failed")
+            return redirect(url_for('Employees'))
     else:
-        flash("Data Updated failed")
-        return redirect(url_for('Employees'))
+        flash("Login to access this page")
+        return redirect(url_for('Loginp'))
 
 
 @app.route('/create-log', methods=['POST'])
